+++ NEWS.md
@@ -1,11 +1,13 @@
 # slider (development version)
 
+* A few edge cases with size zero input in the index functions have been fixed.
+
 * The default for the `.names_to` argument of `*_dfr()` variants has been
   updated to `rlang::zap()` to match the default of the function it is passed
   on to, `vctrs::vec_rbind()`.
 
 * All `*_vec()` variants now maintain size stability when auto-simplifying
-  (i.e. when `.ptype = NULL`) (#78).
+  (i.e. when `.ptype = NULL`) (#78, #93).
 
 * `hop()` and its variants no longer place the names of `.x` on the output.
   Because there is no _size_ guarantee on the output, the size of `.x` can